Initial code

+++ Code/Components/DanpheEMR.ServerModel/BillingModels/BIllingDenominationModel.cs
@@ -13,7 +13,7 @@ namespace DanpheEMR.ServerModel
         public int? DenominationId{ get; set; }
         public int? HandoverId { get; set; }
         public int? CurrencyType { get; set; }
-        public int? Quantity{ get; set; }
+        public double? Quantity{ get; set; }
         public double? Amount { get; set; }
     }
 }
